requirements.txt
flask
openai
twilio
pyngrok
transformers
pinecone-client
langchain
pypdf
tiktoken
tqdm
pandas 
-e .